Refactor: 서버컴포넌트로 변경하다
